adapted test cases

+++ src/main/java/com/ritaja/xchangerate/api/CurrencyConverter.java
@@ -1,8 +1,10 @@
 package com.ritaja.xchangerate.api;
 
+import java.math.BigDecimal;
+
 import com.ritaja.xchangerate.util.Currency;
 
 public interface CurrencyConverter {
 
-	public Double convertCurrency(double moneyAmount, Currency fromCurrency, Currency toCurrency) throws XchangeRateException, CurrencyNotSupportedException;
+	public BigDecimal convertCurrency(BigDecimal moneyAmount, Currency fromCurrency, Currency toCurrency) throws XchangeRateException, CurrencyNotSupportedException;
 }
